LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-MX.html

<!DOCTYPE html><!-- webkit-test-runner [ language=es-MX ] -->
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "es-419");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
